backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "horalix-view-backend"
version = "1.0.0"
description = "Horalix View - Advanced DICOM Viewer Backend with AI Capabilities"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "Horalix Team", email = "dev@horalix.io"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "email-validator>=2.1.0",
    "pydicom>=2.4.0",
    "pylibjpeg>=2.0.0",
    "pylibjpeg-libjpeg>=2.1.0",
    "pylibjpeg-openjpeg>=2.1.0",
    "pylibjpeg-rle>=2.0.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "pillow>=10.2.0",
    "opencv-python-headless>=4.9.0",
    "SimpleITK>=2.3.0",
    "nibabel>=5.2.0",
    "vtk>=9.3.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=3.2.0,<4.0.0",
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    "httpx>=0.26.0",
    "aiofiles>=23.2.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "prometheus-client>=0.19.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "cryptography>=41.0.0",
    "highdicom>=0.22.0",
    "pynetdicom>=2.0.0",
    "dicomweb-client>=0.59.0",
    "fhir.resources>=7.1.0",
]

[project.optional-dependencies]
# AI dependencies - install with: pip install -e ".[ai]"
# NOTE: For CUDA support, install PyTorch separately first:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
# For CPU only:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
ai = [
    # Core ML frameworks
    "torch>=2.2.0",
    "torchvision>=0.17.0",

    # Detection - YOLOv8
    "ultralytics>=8.1.0",

    # Segmentation - MONAI
    "monai>=1.3.0",

    # Interactive Segmentation - SAM/MedSAM
    "segment-anything>=1.0",

    # Supporting libraries
    "einops>=0.7.0",
    "scikit-learn>=1.4.0",
    "timm>=0.9.12",
    "huggingface-hub>=0.20.2",
    "docopt>=0.6.2",
    "psutil>=5.9.0",
    "matplotlib>=3.8.0",
    "scikit-image>=0.22.0",
]

# Extended AI dependencies for additional models
ai-full = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "timm>=0.9.12",
    "huggingface-hub>=0.20.2",
    "tqdm>=4.66.0",
    "docopt>=0.6.2",
    "psutil>=5.9.0",
    "matplotlib>=3.8.0",
    "scikit-image>=0.22.0",
    "monai>=1.3.0",
    "nnunetv2>=2.3.0",
    "segment-anything>=1.0",
    "ultralytics>=8.1.0",
    "transformers>=4.37.0",
    "einops>=0.7.0",
    "diffusers>=0.25.0",
    "accelerate>=0.26.0",
    "scikit-learn>=1.4.0",
    "albumentations>=1.3.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=24.1.0",
    "ruff>=0.1.14",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "httpx>=0.26.0",
    "aiosqlite>=0.19.0",  # For SQLite async tests
]

[project.urls]
Homepage = "https://github.com/horalix/horalix-view"
Documentation = "https://horalix.io/docs"
Repository = "https://github.com/horalix/horalix-view"

[tool.setuptools.packages.find]
where = ["."]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "D", "UP", "S", "B", "A", "C4", "T20", "RET", "SIM"]
ignore = [
    "D100",  # Missing docstring in public module
    "D104",  # Missing docstring in public package
    "D106",  # Missing docstring in public nested class
    "D107",  # Missing docstring in __init__
    "D203",  # 1 blank line required before class docstring (conflicts with D211)
    "D213",  # Multi-line docstring summary should start at the second line (conflicts with D212)
    "D200",  # One-line docstring should fit on one line
    "D212",  # Multi-line docstring summary should start at the first line
    "D400",  # First line should end with a period
    "D415",  # First line should end with a period, question mark, or exclamation point
    "D417",  # Missing argument descriptions in the docstring
    "S101",  # Use of assert detected (allowed in tests)
    "S104",  # Possible binding to all interfaces (intentional for server)
    "S105",  # Possible hardcoded password (false positive for token_type)
    "S106",  # Possible hardcoded password argument (false positive for token_type)
    "B008",  # Do not perform function call in argument defaults (common FastAPI pattern)
    "B904",  # Within an except clause, raise with from (too strict for HTTP exceptions)
    "A002",  # Argument is shadowing a Python builtin (e.g., `format` is common in APIs)
    "RET504",  # Unnecessary assignment before return (style preference)
    "SIM102",  # Use single if instead of nested (can reduce readability)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "S105", "D"]  # Allow assert and skip docstring rules in tests
"app/cli.py" = ["T20", "D401"]  # Allow print statements in CLI and different docstring style
"app/api/**/*.py" = ["B008", "E501", "N803", "N806", "F841", "C401"]  # FastAPI patterns
"app/core/config.py" = ["E501", "T201"]  # Allow longer lines and print for warnings
"app/services/dicom/**/*.py" = ["E501", "N803", "F401", "SIM118", "B007"]  # DICOM code has special requirements
"app/services/ai/**/*.py" = ["E501", "D102", "F841", "SIM118"]  # AI code patterns

[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_ignores = false
disallow_untyped_defs = false
ignore_missing_imports = true
check_untyped_defs = false
strict = false
disable_error_code = ["assignment", "attr-defined", "override", "empty-body", "call-arg", "operator", "var-annotated", "arg-type"]
exclude = [
    "tests/",
    "alembic/",
]

[[tool.mypy.overrides]]
module = [
    "structlog.*",
    "pydicom.*",
    "pynetdicom.*",
    "numpy.*",
    "torch.*",
    "cv2.*",
    "monai.*",
    "ultralytics.*",
    "segment_anything.*",
    "passlib.*",
    "jose.*",
    "cryptography.*",
    "aiofiles.*",
    "SimpleITK.*",
    "nibabel.*",
    "scipy.*",
    "vtk.*",
    "PIL.*",
    "dicomweb_client.*",
    "fhir.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=app --cov-report=term-missing"
